requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

chromadb==0.5.5
jira==3.8.0
pyautogen==0.2.33
python-dotenv==1.0.1
tiktoken==0.7.0
unstructured==0.15.9
